# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Traefik Manager API
  description: REST API for managing Traefik proxy hosts and middleware rules dynamically via a file provider.
  version: "0.2"
servers:
  - url: /api/v1
paths:
...
  /middleware:
    get:
      summary: List Middleware
      responses:
        "200":
          description: A list of custom middlewares
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Middleware'
    post:
      summary: Create Middleware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMiddleware'
      responses:
        "201":
          description: Created middleware
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Middleware'
  /middleware/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    put:
      summary: Update Middleware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMiddleware'
      responses:
        "200":
          description: Updated middleware
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Middleware'
    delete:
      summary: Delete Middleware
      responses:
        "204":
          description: Middleware deleted successfully
...
components:
  schemas:
...
    Middleware:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        name:
          type: string
        type:
          type: string
          enum: [basic_auth, ip_allowlist, rate_limit, headers]
        config:
          type: object
          additionalProperties: true
    CreateMiddleware:
      type: object
      required:
        - name
        - type
        - config
      properties:
        name:
          type: string
        type:
          type: string
          enum: [basic_auth, ip_allowlist, rate_limit, headers]
        config:
          type: object
          additionalProperties: true
...